My final project for my DS course at Insper
This project consists of a project which attempted to use regressors in order
to try and solve classification problems, and compare their performance to
more commonly used classifiers. The results were quite surprising!
The random forest regressor was chosen as the most efficient one, and after
tuning it, an accuracy of 50% (6 possible dimensions) and a mean squared error
of 0,8 was possible to be achieved.